Worklist improvements
