+++ galasa-parent/dev.galasa.framework.api.webui/src/main/java/dev/galasa/framework/api/webui/WorklistQuery.java
@@ -2,8 +2,10 @@ package dev.galasa.framework.api.webui;
 
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -13,18 +15,26 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.component.annotations.ServiceScope;
 
 import com.google.gson.Gson;
+import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 
 import dev.galasa.framework.spi.IFramework;
+import dev.galasa.framework.spi.IResultArchiveStoreDirectoryService;
+import dev.galasa.framework.spi.IRunResult;
+import dev.galasa.framework.spi.ResultArchiveStoreException;
 import dev.galasa.framework.spi.utils.GalasaGsonBuilder;
+import dev.galasa.JsonError;
+import dev.galasa.framework.spi.DssAdd;
+import dev.galasa.framework.spi.DssSwap;
 import dev.galasa.framework.spi.DynamicStatusStoreException;
+import dev.galasa.framework.spi.IDynamicStatusStoreService;
 
 
 @Component(service = Servlet.class, scope = ServiceScope.PROTOTYPE, property =
@@ -37,14 +47,12 @@ public class WorklistQuery extends HttpServlet {
 	IFramework framework;
 	
 	final static Gson gson = GalasaGsonBuilder.build();
-    
-    private WorklistUtils worklistUtils;
 
 	
 	@Override 
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		
-		String principalUsername = worklistUtils.getPrincipalUsername(req);
+		String principalUsername = getPrincipalUsername(req);
 		
 		returnWorklist(principalUsername, resp, 200);
 
@@ -54,24 +62,26 @@ public class WorklistQuery extends HttpServlet {
 	@Override
 	protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		
-		String principalUsername = worklistUtils.getPrincipalUsername(req);
+		String principalUsername = getPrincipalUsername(req);
 		
 		Map<String, String> paramMap = getParameterMap(req);
 	
 		String runId = "";
-		if (paramMap.get("runId") != null && !paramMap.get("runId").equals("")) {
+		if (paramMap.get("runId") != null && !paramMap.get("runId").isEmpty()) {
 			runId = paramMap.get("runId");
+		} else {
+			sendJsonError(resp);
 		}
 		
-
+		boolean newResourceCreated = false;
 		try {
-			worklistUtils.addRunIdToDss(principalUsername, runId);
+			newResourceCreated = addRunIdToDss(principalUsername, runId);
 		} catch (Exception e) {
 			throw new ServletException("Error while updating the DSS", e);
 		}
 		
 		int statusCode = 200;
-		if (worklistUtils.newResourceCreated == true) {
+		if (newResourceCreated == true) {
 			statusCode = 201;
 		}
 		
@@ -83,18 +93,19 @@ public class WorklistQuery extends HttpServlet {
 	@Override
 	protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		
-		String principalUsername = worklistUtils.getPrincipalUsername(req);
+		String principalUsername = getPrincipalUsername(req);
 		
 		Map<String, String> paramMap = getParameterMap(req);
 		
 		String runId = "";
-		if (paramMap.get("runId") != null && !paramMap.get("runId").equals("")) {
+		if (paramMap.get("runId") != null && !paramMap.get("runId").isEmpty()) {
 			runId = paramMap.get("runId");
+		} else {
+			sendJsonError(resp);
 		}
-		
 
 		try {
-			worklistUtils.removeRunIdFromDss(principalUsername, runId);
+			removeRunIdFromDss(principalUsername, runId);
 		} catch (Exception e) {
 			throw new ServletException("Error while updating the DSS", e);
 		}
@@ -104,18 +115,18 @@ public class WorklistQuery extends HttpServlet {
 	}
 	
 	
-	protected void returnWorklist(String principalUsername, HttpServletResponse resp, int statusCode) throws ServletException {
+	private void returnWorklist(String principalUsername, HttpServletResponse resp, int statusCode) throws ServletException {
 		
 		List<String> runIdsInWorklist = new ArrayList<String>();
 		
 		try {
-			 runIdsInWorklist = worklistUtils.getRunIdsFromDss(principalUsername);
+			 runIdsInWorklist = getRunIdsFromDss(principalUsername);
 		} catch (DynamicStatusStoreException e) {
 			throw new ServletException("Error while retrieving Worklist property from the DSS", e);
 		}
 		
 
-		List<WorklistItem> runs = worklistUtils.getWorklistItems(principalUsername, runIdsInWorklist);
+		List<WorklistItem> runs = getWorklistItems(principalUsername, runIdsInWorklist);
 		
 		
 		JsonObject worklist = new JsonObject();
@@ -125,15 +136,13 @@ public class WorklistQuery extends HttpServlet {
 		
 		String json = "";
 		 
-		if (!worklist.isJsonNull()) {
-			try {
-				json = gson.toJson(worklist);
-			} catch (Exception e) {
-				throw new ServletException("An error occurred whilst retrieving the Worklist", e);
-			}
-		 }
+		try {
+			json = gson.toJson(worklist);
+		} catch (Exception e) {
+			throw new ServletException("An error occurred whilst retrieving the Worklist", e);
+		}
 		
-		 try {
+		try {
 	         PrintWriter out = resp.getWriter();
 	         resp.setContentType( "Application/json");
 	         resp.addHeader("Access-Control-Allow-Origin", "*");
@@ -146,7 +155,23 @@ public class WorklistQuery extends HttpServlet {
 		 
 		
 	}
-
+	
+	private String getPrincipalUsername(HttpServletRequest req) {
+		
+		Principal pU = req.getUserPrincipal();
+		
+		String principalUsername = "";
+		
+		if (pU == null) {
+			principalUsername = "unknown";
+		} else {
+			principalUsername = pU.toString().toLowerCase();
+		}
+		
+		return principalUsername;
+		
+	}
+	
 	
 	private Map<String, String> getParameterMap(HttpServletRequest request) {
 
@@ -165,9 +190,252 @@ public class WorklistQuery extends HttpServlet {
 	}
 	
 	
-	@Activate
-	public void activate() {
-	   this.worklistUtils = new WorklistUtils(this.framework);
+	private List<String> getRunIdsFromDss(String principalUsername) throws DynamicStatusStoreException, ServletException {
+		
+		
+		IDynamicStatusStoreService dss = framework.getDynamicStatusStoreService("webui");
+		String worklistProperty = dss.get("user." + principalUsername + ".worklist");
+		
+		if (worklistProperty != null) {
+			
+			List<String> runIdsInWorklist = new ArrayList<String>();	
+			
+			if (worklistProperty.length() != 0) {
+				
+				JsonObject jsonObject = new JsonParser().parse(worklistProperty).getAsJsonObject();
+				
+				JsonArray runIds = (JsonArray) jsonObject.get("runIds");		
+
+				for (JsonElement el : runIds) {
+					runIdsInWorklist.add(el.getAsString());
+				}	
+			}	
+			
+			return runIdsInWorklist;
+			
+		} else {
+			return null;
+		}
+		
+	}
+	
+	
+	private List<WorklistItem> getWorklistItems(String principalUsername, List<String> runIdsInWorklist) throws ServletException {
+		
+		List<WorklistItem> runs = new ArrayList<WorklistItem>();
+		
+		List<String> runIdsNotFound = new ArrayList<String>();
+		
+		if (runIdsInWorklist != null) {
+			
+			IRunResult run = null;
+			for (String runId : runIdsInWorklist) {
+				try {
+					run = getRunById(runId);
+				} catch (ResultArchiveStoreException e) {
+					throw new ServletException("Error while retrieving run for Run ID " + runId, e);
+				}
+				
+				if (run != null) {
+					
+					try {
+						
+						String runName = run.getTestStructure().getRunName();
+						String shortName = run.getTestStructure().getTestShortName();
+						String testClass = run.getTestStructure().getTestName();
+						String result = run.getTestStructure().getResult();
+						
+						WorklistItem worklistItem = new WorklistItem(runId, runName, shortName, testClass, result);
+						
+						runs.add(worklistItem);
+						
+					} catch (ResultArchiveStoreException e) {
+						throw new ServletException("Error while retriving Worklist information for Run ID " + runId, e);
+					}
+					
+				} else {
+					runIdsNotFound.add(runId);
+					try {
+						removeRunIdFromDss(principalUsername, runId);
+					} catch (Exception e) {
+						throw new ServletException("Error while updating the DSS", e);
+					}
+				}
+				
+			}
+			
+			runIdsInWorklist.removeAll(runIdsNotFound);
+		}
+		
+		return runs;
+	}
+	
+	
+	private IRunResult getRunById(String id) throws ResultArchiveStoreException {
+
+		IRunResult run = null;
+	       
+        for (IResultArchiveStoreDirectoryService directoryService : framework.getResultArchiveStore().getDirectoryServices()) {
+           
+           run = directoryService.getRunById(id);
+           
+           if(run != null) {
+        	   return run;
+           }
+           
+        }
+		
+		return null;
+	}
+	
+	
+	private boolean addRunIdToDss(String principalUsername, String runId) throws DynamicStatusStoreException, ServletException {
+		
+		IDynamicStatusStoreService dss = framework.getDynamicStatusStoreService("webui");
+		
+		boolean exceptionThrown = false;
+		
+		do {
+			
+			String oldValue = dss.get("user." + principalUsername + ".worklist");
+			
+			if (oldValue != null) {
+				
+				JsonObject jsonObject = new JsonParser().parse(oldValue).getAsJsonObject();
+				JsonArray runIds = (JsonArray) jsonObject.get("runIds");
+				
+				for (JsonElement el : runIds) {
+					if (el.getAsString().equals(runId)) {
+						// If Run ID is already in Worklist, exit this method
+						return false;
+					}
+				}
+				
+				runIds.add(runId);			
+				
+				String newValue = jsonObject.toString();
+				
+				
+				int count = 0;
+				
+
+				try {
+					dss.performActions(
+							new DssSwap("user." + principalUsername + ".worklist", oldValue, newValue));
+					exceptionThrown = false;
+				} catch (DynamicStatusStoreException e) {
+					exceptionThrown = true;
+					oldValue = dss.get("user." + principalUsername + ".worklist");
+					count++;
+					
+					if (count > 1000) {
+						throw new ServletException("Error while updating the DSS, maximum attempts reached");
+					}
+				}
+					
+				return false;
+			
+				
+			} else {
+				
+				createWorklistPropertyInDss(principalUsername, runId);
+				
+				return true;
+			}
+			
+		} while (exceptionThrown);
+			
+	}
+	
+	
+	protected void removeRunIdFromDss(String principalUsername, String runId) throws DynamicStatusStoreException, ServletException {
+		
+		IDynamicStatusStoreService dss = framework.getDynamicStatusStoreService("webui");
+		
+		boolean exceptionThrown = false;
+		
+		do {
+		
+			String oldValue = dss.get("user." + principalUsername + ".worklist");
+			
+			if (oldValue != null) {
+				
+				JsonObject jsonObject = new JsonParser().parse(oldValue).getAsJsonObject();
+				JsonArray runIds = (JsonArray) jsonObject.get("runIds");
+				
+				Iterator<JsonElement> it = runIds.iterator();
+				
+				boolean found = false;
+				while (it.hasNext()) {
+					if (it.next().getAsString().equals(runId)){
+						it.remove();
+						found = true;
+					}
+				}
+				
+				if (found == false) {
+					return;
+				}
+			
+				String newValue = jsonObject.toString();
+			
+				int count = 0;
+				
+				try {
+					dss.performActions(
+							new DssSwap("user." + principalUsername + ".worklist", oldValue, newValue));
+					exceptionThrown = false;
+				} catch (DynamicStatusStoreException e) {
+					exceptionThrown = true;
+					oldValue = dss.get("user." + principalUsername + ".worklist");
+					count++;
+					
+					if (count > 1000) {
+						throw new ServletException("Error while updating the DSS, maximum attempts reached");
+					}
+				}
+					
+			} 
+		
+		} while (exceptionThrown);
+				
+	}
+
+	
+	private void createWorklistPropertyInDss(String principalUsername, String runId) throws DynamicStatusStoreException, ServletException {
+		
+		IDynamicStatusStoreService dss = framework.getDynamicStatusStoreService("webui");
+		
+		JsonObject jsonObject = new JsonObject();
+		JsonArray runIds = new JsonArray();
+		runIds.add(runId);
+		jsonObject.add("runIds", runIds);
+		
+		String value = jsonObject.toString();
+		
+		try {
+			dss.performActions(
+					new DssAdd("user." + principalUsername + ".worklist", value));
+		} catch (DynamicStatusStoreException e) {
+			// If the property already exists at the time of adding, attempt DssSwap instead
+			addRunIdToDss(principalUsername, runId);
+			return;
+		}
+		
+	}
+	
+	
+	private void sendJsonError(HttpServletResponse resp) throws IOException {
+		
+		PrintWriter out = resp.getWriter();
+		resp.setStatus(400);
+		resp.setContentType("Application/json");
+        resp.addHeader("Access-Control-Allow-Origin", "*");
+        JsonError error = new JsonError("Run ID missing from query");
+        String jsonError = gson.toJson(error);
+        out.print(jsonError);
+        out.close();
+		
 	}
 
 }
